refactor(ui/output): explicit __init__, consolidate timing helpers, add Protocol
Three follow-up cleanups after the UI module review:

- output/__init__.py: replace star-import-from-output.py with direct imports
  from each source module and an explicit __all__. Delete output.py (it was a
  pure re-export middleman with no logic of its own).

- time_format.py: add _fmt_timing and _elapsed_hms, which were duplicated in
  output/labels.py. Update labels.py and the six output/ modules that imported
  from labels (live_display, renderers, repl_display, tool_details,
  tool_tracking, tracker) to import from time_format directly — one
  canonical location for all duration/elapsed formatting.

- tool_tracking.py: add ToolTrackingSupport Protocol so type-checkers and
  IDEs can validate that concrete implementers satisfy the mixin's contract
  rather than discovering the gap at runtime via NotImplementedError.
