+++ app/cli/interactive_shell/ui/output/renderers.py
@@ -11,7 +11,8 @@ from app.cli.interactive_shell.ui.output.environment import (
     _safe_print,
     get_output_format,
 )
-from app.cli.interactive_shell.ui.output.labels import BADGE_STYLES, _elapsed_hms
+from app.cli.interactive_shell.ui.output.labels import BADGE_STYLES
+from app.cli.interactive_shell.ui.time_format import _elapsed_hms
 from app.cli.interactive_shell.ui.theme import (
     BRAND,
     DIM,
