+++ app/cli/interactive_shell/ui/output/tool_details.py
@@ -4,7 +4,7 @@ from typing import Any
 
 from rich.text import Text
 
-from app.cli.interactive_shell.ui.output.labels import _elapsed_hms, _fmt_timing
+from app.cli.interactive_shell.ui.time_format import _elapsed_hms, _fmt_timing
 from app.cli.interactive_shell.ui.theme import BRAND, DIM, HIGHLIGHT, SECONDARY, TEXT
 from app.tools.registry import get_registered_tool_map, resolve_tool_display_name
 from app.utils.tool_trace import format_json_preview
