+++ app/cli/interactive_shell/ui/time_format.py
@@ -5,6 +5,18 @@ from __future__ import annotations
 from datetime import UTC, datetime
 from typing import Literal
 
+
+def _fmt_timing(elapsed_ms: int) -> str:
+    """Format a millisecond count as a short elapsed string: ``42ms`` or ``1.3s``."""
+    return f"{elapsed_ms / 1000:.1f}s" if elapsed_ms >= 1000 else f"{elapsed_ms}ms"
+
+
+def _elapsed_hms(seconds: float) -> str:
+    """Format seconds as ``M:SS`` for live-display rows."""
+    m = int(seconds // 60)
+    s = int(seconds % 60)
+    return f"{m}:{s:02d}"
+
 ReplTimestampStyle = Literal["table", "compact", "utc"]
